Added logger

+++ src/main/java/com/campaign/demo/controller/CampaignController.java
@@ -3,6 +3,8 @@ package com.campaign.demo.controller;
 import com.campaign.demo.entity.Campaign;
 import com.campaign.demo.error.ResourceNotFoundException;
 import com.campaign.demo.service.CampaignService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,6 +17,7 @@ import java.util.Optional;
 @RestController
 @RequestMapping("/demo")
 public class CampaignController {
+    public static final Logger logger = LoggerFactory.getLogger(CampaignController.class);
 
     @Autowired
     private CampaignService campaignService;
@@ -22,25 +25,27 @@ public class CampaignController {
     @GetMapping(value = "/campaigns")
     public ResponseEntity<List<Campaign>> getCampaigns() {
         List<Campaign> campaigns = campaignService.getCampaigns();
+        if (campaigns.isEmpty()) {
+            return new ResponseEntity(HttpStatus.NO_CONTENT);
+        }
 
-        ResponseEntity<List<Campaign>> response = new ResponseEntity<>(campaigns, HttpStatus.OK);
-
-        return response;
+        return new ResponseEntity<>(campaigns, HttpStatus.OK);
     }
 
     @PostMapping(value = "/campaign/create")
     public ResponseEntity<Campaign> createCampaign(@Valid @RequestBody Campaign campaign) {
         Campaign createCampaign = campaignService.saveCampaign(campaign);
 
-        ResponseEntity<Campaign> response = new ResponseEntity<>(createCampaign, HttpStatus.OK);
-
-        return response;
+        return new ResponseEntity<>(createCampaign, HttpStatus.OK);
     }
 
     @PutMapping(value = "/campaign/update")
     public ResponseEntity<Campaign> updateCampaign(@RequestParam("id") Integer campaignId, @Valid @RequestBody Campaign campaignDetails) {
         Optional<Campaign> campaign = campaignService.getCampaign(campaignId);
-        campaign.orElseThrow(() -> new ResourceNotFoundException("Campaign", "id", campaignId));
+        if (!campaign.isPresent()) {
+            logger.error("Unable to update. Campaign not found with id: " + campaignId);
+            throw new ResourceNotFoundException("Campaign", "id", campaignId);
+        }
 
         Campaign currentCampaign = campaign.get();
         currentCampaign.setDiscountType(campaignDetails.getDiscountType());
@@ -61,7 +66,10 @@ public class CampaignController {
     @DeleteMapping(value = "/campaign/delete")
     public ResponseEntity<Campaign> deleteCampaign(@RequestParam("id") Integer campaignId) {
         Optional<Campaign> campaign = campaignService.getCampaign(campaignId);
-        campaign.orElseThrow(() -> new ResourceNotFoundException("Campaign", "id", campaignId));
+        if (!campaign.isPresent()) {
+            logger.error("Unable to delete. Campaign not found with id: " + campaignId);
+            throw new ResourceNotFoundException("Campaign", "id", campaignId);
+        }
 
         campaignService.deleteCampaign(campaignId);
